+++ src/CameraServiceProvider.php
@@ -12,6 +12,7 @@ use TromsFylkestrafikk\Camera\Console\CameraSet;
 use TromsFylkestrafikk\Camera\Console\CameraShow;
 use TromsFylkestrafikk\Camera\Console\FolderWatcher;
 use TromsFylkestrafikk\Camera\Console\FindLatest;
+use TromsFylkestrafikk\Camera\Jobs\DetectStale;
 use TromsFylkestrafikk\Camera\Services\CameraTokenizer;
 
 class CameraServiceProvider extends ServiceProvider
@@ -70,14 +71,14 @@ class CameraServiceProvider extends ServiceProvider
 
     protected function registerScheduledCommands()
     {
-        $scanPeriod = config('camera.poor_mans_inotify');
-        if (!$scanPeriod) {
-            return;
-        }
-        $this->app->booted(function () use ($scanPeriod) {
+        $this->app->booted(function () {
             // @var \Illuminate\Console\Scheduling\Schedule $schedule
             $schedule = $this->app->make(Schedule::class);
-            $schedule->command(FindLatest::class)->cron(sprintf("*/%d * * * *", $scanPeriod));
+            $scanPeriod = config('camera.poor_mans_inotify');
+            if ($scanPeriod) {
+                $schedule->command(FindLatest::class)->cron(sprintf("*/%d * * * *", $scanPeriod));
+            }
+            $schedule->job(DetectStale::class)->everyMinute();
         });
     }