+++ src/Console/CameraList.php
@@ -38,9 +38,14 @@ class CameraList extends Command
      */
     public function handle()
     {
-        $cameras = Camera::select(['id', 'camera_id', 'name', 'updated_at'])->get();
+        $cameras = Camera::select(['id', 'camera_id', 'name', 'currentFile', 'currentDate', 'active'])
+            ->get()
+            ->map(function ($camera) {
+                $camera->setAppends([]);
+                return $camera;
+            });
         $this->table(
-            ['ID', 'Cam ID', 'Name', 'Last seen'],
+            ['ID', 'Cam ID', 'Name', 'Current File', 'Received', 'Live'],
             $cameras
         );
         return 0;