+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 - Package extracted from other project
-- API call for getting the latest image, actual image, not URL to its file.
 - Artisan command for watching camera folder(s) which broadcasts
   updated/new images when dropped there.
 - Camera model contains current/latest image.
+- API call for camera model returns image, URL to it, mime and timestamp.
 - Add optional scheduled check for latest images. Useful when inotify
   isn't available.
+- Cron job looks for stalled cameras and deactivates them if latest
+  image is older than the `max_age` configuration option.